Issue #21: Refactor auth.js to authSignup.js and auth.test.js to authSignupController.test.js for clarity
