+++ routes/users.js
@@ -2,7 +2,7 @@ var express = require('express');
 var router = express.Router();
 const bcrypt = require('bcryptjs');
 var passport = require('passport');
-const { authSignupController } = require('../controllers/auth');
+const { authSignupController } = require('../authSignupController/authSignup');
 const User = require('../models/User');
 const mongoose = require('mongoose');
 var db = require('../config/keys').MongoURI;
